+++ noodles-cram/src/reader.rs
@@ -339,14 +339,10 @@ where
         &'a mut self,
         reference_sequence_repository: &'a fasta::Repository,
         header: &'a sam::Header,
-    ) -> Box<dyn Iterator<Item = io::Result<Box<dyn sam::AlignmentRecord>>> + 'a> {
+    ) -> Box<dyn Iterator<Item = io::Result<sam::alignment::Record>> + 'a> {
         Box::new(
             self.records(reference_sequence_repository, header)
-                .map(|result| {
-                    result
-                        .and_then(|record| record.try_into_alignment_record(header))
-                        .map(|record| Box::new(record) as Box<dyn sam::AlignmentRecord>)
-                }),
+                .map(|result| result.and_then(|record| record.try_into_alignment_record(header))),
         )
     }
 }