+++ noodles-sam/src/reader.rs
@@ -10,8 +10,7 @@ use std::io::{self, BufRead, Read, Seek};
 use noodles_bgzf as bgzf;
 use noodles_fasta as fasta;
 
-use super::{AlignmentReader, AlignmentRecord, Header};
-use crate::{alignment::Record, lazy};
+use super::{alignment::Record, lazy, AlignmentReader, Header};
 
 /// A SAM reader.
 ///
@@ -278,11 +277,8 @@ where
         &'a mut self,
         _: &'a fasta::Repository,
         header: &'a Header,
-    ) -> Box<dyn Iterator<Item = io::Result<Box<dyn AlignmentRecord>>> + 'a> {
-        Box::new(
-            self.records(header)
-                .map(|result| result.map(|record| Box::new(record) as Box<dyn AlignmentRecord>)),
-        )
+    ) -> Box<dyn Iterator<Item = io::Result<Record>> + 'a> {
+        Box::new(self.records(header))
     }
 }